nixos-render-docs: move docbook renderer to docbook module
